Move GetPortURL to ingress e2e fw
